+++ benchmark/scripts/run.php
@@ -1,38 +1,32 @@
 <?php
 
 $opt = getopt("", array(
-	/** @var string $engine */
-	"engine:",
-	/** @var string $template */
-	"template:",
-	/** @var string $data */
-	"data:",
-	/** @var boolean $double */
-	"double",
-	/** @var boolean $stress */
-	"stress",
-	/** @var string $message */
-	"message:"
+    /** @var string $engine */
+    "engine:",
+    /** @var string $template */
+    "template:",
+    /** @var string $data */
+    "data:",
+    /** @var boolean $double */
+    "double",
+    /** @var string $message */
+    "message:",
+    /** @var boolean $stress */
+    "stress:",
+    /** @var boolean $auto_reload */
+    "auto_reload"
 ));
 
 require_once __DIR__.'/bootstrap.php';
 
+$opt += array(
+    "message"     => "plain",
+    "stress"      => 0,
+);
+
 extract($opt);
 
-if (isset($stress)) {
-	$start = microtime(true);
-	$message = 'stress test';
-	$data = json_decode(file_get_contents($data), true);
-	gc_enable();
-	for ($i = 0; $i < Benchmark::STRESS_REQUEST_COUNT; $i++) {
-		Benchmark::$engine($template, $data, false);
-		if ($i % 50 == 0) gc_collect_cycles();
-	}
-	$time = microtime(true) - $start;
 
-} else {
-	$time = Benchmark::$engine($template, json_decode(file_get_contents($data), true), isset($double));
+$time = Benchmark::$engine($template, json_decode(file_get_contents($data), true), isset($double), $stress, isset($auto_reload));
 
-}
 printf(Benchmark::OUTPUT, $engine, $message, round($time, 4), round(memory_get_peak_usage()/1024/1024, 2));
-