+++ tests/cases/Fenom/AutoEscapeTest.php
@@ -13,27 +13,35 @@ class AutoEscapeTest extends TestCase {
         );
         return array(
             // variable
-            array('{$html}', $html, $vars, 0),
-            array('{$html}', $escaped, $vars, \Fenom::AUTO_ESCAPE),
-            array('{raw $html}', $html, $vars, \Fenom::AUTO_ESCAPE),
+            array('{$html}, {$html}', "$html, $html", $vars, 0),
+            array('{$html}, {$html}', "$escaped, $escaped", $vars, \Fenom::AUTO_ESCAPE),
             array('{raw $html}, {$html}', "$html, $escaped", $vars, \Fenom::AUTO_ESCAPE),
-            array('{raw "{$html|up}"}', strtoupper($html), $vars, \Fenom::AUTO_ESCAPE),
+            array('{raw $html}, {$html}', "$html, $escaped", $vars, \Fenom::AUTO_ESCAPE),
+            array('{raw "{$html|up}"}, {$html}', strtoupper($html).", $escaped", $vars, \Fenom::AUTO_ESCAPE),
             array('{autoescape true}{$html}{/autoescape}, {$html}', "$escaped, $html", $vars, 0),
             array('{autoescape false}{$html}{/autoescape}, {$html}', "$html, $escaped", $vars, \Fenom::AUTO_ESCAPE),
             array('{autoescape true}{$html}{/autoescape}, {$html}', "$escaped, $escaped", $vars, \Fenom::AUTO_ESCAPE),
             array('{autoescape false}{$html}{/autoescape}, {$html}', "$html, $html", $vars, 0),
+            array('{autoescape true}{raw $html}{/autoescape}, {$html}', "$html, $html", $vars, 0),
+            array('{autoescape false}{raw $html}{/autoescape}, {$html}', "$html, $escaped", $vars, \Fenom::AUTO_ESCAPE),
+            array('{autoescape true}{raw $html}{/autoescape}, {$html}', "$html, $escaped", $vars, \Fenom::AUTO_ESCAPE),
+            array('{autoescape false}{raw $html}{/autoescape}, {$html}', "$html, $html", $vars, 0),
 
             // inline function
-            array('{test_function text=$html}', $html, $vars, 0),
-            array('{test_function text=$html}', $escaped, $vars, \Fenom::AUTO_ESCAPE),
-            array('{raw:test_function text=$html}', $html, $vars, \Fenom::AUTO_ESCAPE),
-            array('{raw:test_function text="{$html|up}"}', strtoupper($html), $vars, \Fenom::AUTO_ESCAPE),
+            array('{test_function text=$html}, {$html}', "$html, $html", $vars, 0),
+            array('{test_function text=$html}, {$html}', "$escaped, $escaped", $vars, \Fenom::AUTO_ESCAPE),
+            array('{raw:test_function text=$html}, {$html}', "$html, $escaped", $vars, \Fenom::AUTO_ESCAPE),
+            array('{raw:test_function text="{$html|up}"}, {$html}', strtoupper($html).", $escaped", $vars, \Fenom::AUTO_ESCAPE),
             array('{autoescape true}{test_function text=$html}{/autoescape}, {test_function text=$html}', "$escaped, $html", $vars, 0),
             array('{autoescape false}{test_function text=$html}{/autoescape}, {test_function text=$html}', "$html, $escaped", $vars, \Fenom::AUTO_ESCAPE),
             array('{autoescape true}{test_function text=$html}{/autoescape}, {test_function text=$html}', "$escaped, $escaped", $vars, \Fenom::AUTO_ESCAPE),
             array('{autoescape false}{test_function text=$html}{/autoescape}, {test_function text=$html}', "$html, $html", $vars, 0),
+            array('{autoescape true}{raw:test_function text=$html}{/autoescape}, {test_function text=$html}', "$html, $html", $vars, 0),
+            array('{autoescape false}{raw:test_function text=$html}{/autoescape}, {test_function text=$html}', "$html, $escaped", $vars, \Fenom::AUTO_ESCAPE),
+            array('{autoescape true}{raw:test_function text=$html}{/autoescape}, {test_function text=$html}', "$html, $escaped", $vars, \Fenom::AUTO_ESCAPE),
+            array('{autoescape false}{raw:test_function text=$html}{/autoescape}, {test_function text=$html}', "$html, $html", $vars, 0),
 
-            // block function. Has bug, disable for vacation
+            // block function. Have bugs
 //            array('{test_block_function}{$html}{/test_block_function}', $html, $vars, 0),
 //            array('{test_block_function}{$html}{/test_block_function}', $escaped, $vars, \Fenom::AUTO_ESCAPE),
 //            array('{raw:test_block_function}{$html}{/test_block_function}', $html, $vars, \Fenom::AUTO_ESCAPE),