+++ CHANGELOG.md
@@ -1,6 +1,39 @@
 CHANGELOG
 =========
 
+## 1.1.0
+
+- Bug #19: Bug with "if" expressions starting with "("
+- Bug #16: Allow modifiers for function calls
+- Bug #25: Invalid option flag for `auto_reload`
+- Bug: Invalid options for cached templates
+- Bug: Removed memory leak after render
+- Fix nested bracket pull #10
+- Fix bugs with provider
+- Improve providers' performance
+- Improve #1: Add `is` and `in` operator
+- Remove Fenom::addTemplate(). Use providers for adding custom templates.
+- Big refractory: parsers, providers, storage
+- Improve tokenizer
+- Internal optimization
+- Add options for benchmark
+- Add stress test (thanks to @klkvsk)
+- Bugs--
+- Comments++
+- Docs++
+- Test++
+
+## 1.0.8
+
+- Perform auto_escape options
+- Fix bugs
+- Update documentation
+
+## 1.0.7
+
+- Perform auto_escape options
+- Fix bugs
+
 ## 1.0.6 (2013-07-04)
 
 - Fix modifiers insertions